Refactoring DataManager.java into smaller helper classes that are Spring friendly
