+++ core/src/main/java/org/fao/geonet/kernel/IndexMetadataTask.java
@@ -43,7 +43,7 @@ import javax.annotation.Nullable;
 /**
  * A runnable for indexing multiple metadata in a separate thread.
  */
-final class IndexMetadataTask implements Runnable {
+public final class IndexMetadataTask implements Runnable {
 
     private final ServiceContext _context;
     private final List<?> _metadataIds;
@@ -60,7 +60,7 @@ final class IndexMetadataTask implements Runnable {
      * @param metadataIds       the metadata ids to index (either integers or strings)
      * @param transactionStatus if non-null, wait for the transaction to complete before indexing
      */
-    IndexMetadataTask(@Nonnull ServiceContext context, @Nonnull List<?> metadataIds, Set<IndexMetadataTask> batchIndex,
+    public IndexMetadataTask(@Nonnull ServiceContext context, @Nonnull List<?> metadataIds, Set<IndexMetadataTask> batchIndex,
                       @Nullable TransactionStatus transactionStatus, @Nonnull AtomicInteger indexed) {
         this.indexed = indexed;
         this._transactionStatus = transactionStatus;
